created ScormEvent class and refactored scorm events
