+++ src/Controller.php
@@ -22,7 +22,7 @@ class Controller extends PhpObj {
         'training_session_attend' => 'Attended',
         'training_session_enrol' => 'EventEnrol',
         'training_session_unenrol' => 'EventUnenrol',
-        'scorm_scoreraw_submitted' => 'ScormScorerawSubmitted',
+        'scorm_scoreraw_submitted' => 'ScormScoreRawSubmitted',
         'scorm_status_submitted' => 'ScormStatusSubmitted'
     ];
